Sentinel: Add details on updated provider_name field for TF 0.13
Starting with Terraform 0.13, the provider_name field in the resource
collections in the tfplan/v2 and tfstate/v2 imports will contain the
fully qualified path to the resource in the Terraform Registry. Example:
registry.terraform.io/hashicorp/null for the null provider.

This change adds a note for this in the tfplan/v2 and tfstate/v2
documentation pages.